Installing the SMF manifest to a directory in the install prefix
I realize this is potentially controversial.  The issue I ran into was
that the `install-illumos` command failed on SmartOS because
`/lib` is not writable.  Given that there's no `svccfg import` happening
in here, I'm hoping the new location is OK, although perhaps it would be
less problematic if there *were* an `svccfg import` :confused: