Want runstate dir for plugins
Prior to v2.0.0 there was a runstate directory under the install
prefix that was used for keeping local state, specifically the
"circpkg" plugin, used internally at Circonus. It is used to keep
a cache of output from a relatively expensive operation, and to
allow for system-local tuning by an operator.

This change restores that concept though at a slightly different
location, to jive with the current default prefix.